pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "screenshot-saas"
version = "0.1.0"
description = "A SaaS application for extracting screenshots from app stores"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.4",
    "uvicorn[standard]>=0.34.0",
    "playwright>=1.57.0",
    "aiohttp>=3.11.10",
    "beautifulsoup4>=4.12.3",
    "redis>=5.2.0",
    "celery>=5.4.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.7.0",
    "boto3>=1.35.82",
    "stripe>=11.3.0",
    "sentry-sdk>=2.20.0",
    "psycopg2-binary>=2.9.10",
    "python-multipart>=0.0.20",
    "python-slugify>=8.0.4",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "pre-commit>=3.0"
]
test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true